+++ packages/workspace-tools/CHANGELOG.md
@@ -1,9 +1,9 @@
 ## 1.182.1 (2024-09-11)
 
-
 ### Bug Fixes
 
-- **workspace-tools:** Resolve issue with input entry point names ([12ca61d3](https://github.com/storm-software/storm-ops/commit/12ca61d3))
+- **workspace-tools:** Resolve issue with input entry point names
+  ([12ca61d3](https://github.com/storm-software/storm-ops/commit/12ca61d3))
 
 ## 1.182.0 (2024-09-11)